updated decorator to include tags

+++ care/emr/api/viewsets/facility.py
@@ -114,23 +114,22 @@ class FacilityViewSet(EMRModelViewSet):
             raise PermissionDenied("Only Super Admins can delete Facilities")
 
     @method_decorator(parser_classes([MultiPartParser]))
-    @action(methods=["POST"], detail=True)
+    @action(methods=["POST", "DELETE"], detail=True)
     def cover_image(self, request, external_id):
         facility = self.get_object()
         self.authorize_update({}, facility)
-        serializer = FacilityImageUploadSerializer(facility, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data)
 
-    @cover_image.mapping.delete
-    def cover_image_delete(self, *args, **kwargs):
-        facility = self.get_object()
-        self.authorize_update({}, facility)
-        delete_cover_image(facility.cover_image_url, "cover_images")
-        facility.cover_image_url = None
-        facility.save()
-        return Response(status=204)
+        if request.method == "POST":
+            serializer = FacilityImageUploadSerializer(facility, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(serializer.data)
+        if request.method == "DELETE":
+            delete_cover_image(facility.cover_image_url, "cover_images")
+            facility.cover_image_url = None
+            facility.save()
+            return Response(status=204)
+        return Response(data="Method Not Allowed", status=405)
 
 
 @generate_swagger_schema_decorator
